some slight changes
please use this format for the other functions. you don't have to implement the functions, just make them void(don't pass to them anything and don't get from them anything) and that's all. we will do the functions later. we just need from you to declare the function and that's all.